requirements.txt
# Data processing
numpy
pandas
scikit-learn

# Image handling
Pillow

# Visualization
matplotlib
seaborn
tqdm

# To access .env
python-dotenv

# streamlit
streamlit

# Deep learning
torch
torchvision

# ONNX export and runtime
# onnx
# onnxruntime

# Deployment
flask
boto3
docker

# Jupyter / notebooks
jupyter
ipykernel
